@@ -1,5 +1,5 @@
 module github.com/libdns/cloudflare
 
-go 1.14
+go 1.18
 
-require github.com/libdns/libdns v0.2.0
+require github.com/libdns/libdns v0.2.2-0.20221006221142-3ef90aee33fd
